Add files via upload
A chat-bots aims to make a conversation between both human and machine. The machine has been embedded knowledge to identify the sentences and making a decision itself as a response to answer a question. Chat-bots will be completely based on a text-based user interface, allowing the user to type commands and receive text as well as text to speech response. Chat-bots are usually stateful services, remembering previous commands in order to provide functionality. It can be utilized securely by an even larger audience when chat-bots technology is integrated with popular web services. The college inquiry chat-bots will be built using artificial algorithms that analyze user's queries and understand user's message. The response principle is matching the input sentence from a user. The User can ask the question any college-related activities through the chat-bot without physically available to the college for inquiry. The System analyses the question and then answers to the user. With the help of artificial intelligence, the system answers the query asked by the students. The system replies using an effective Graphical User Interface as if a real person is talking to the user. The user just has to register himself to the system and has to login to the system. The chat-bots consists of core and interface that is accessing the core in (MySQL).Natural language processing technologies are used for parsing, tokenizing, stemming and filtering the content of the chats. 

Front End: Python 

Back End: MySQL